Renamed gateway to reflect Omnipay convention
